server/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.34.0
pydantic==2.10.0
pydantic-settings==2.7.0

# Database
motor==3.6.0
pymongo==4.9.2
redis==5.0.3

# AI
google-generativeai==0.8.0

# RAG
chromadb==0.6.3
PyMuPDF==1.25.0

# Embeddings (local, zero cost)
sentence-transformers==3.3.0

# Scaping/Web (Legal Base)
beautifulsoup4==4.13.0
requests==2.32.0
httpx==0.28.1
aiohttp==3.11.10

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
bcrypt==4.2.0
PyJWT==2.10.0
cachetools==5.3.3

# Document Export
python-docx==1.2.0
markdown2==2.5.4
weasyprint==62.3
rank_bm25==0.2.2

# RPA & Browser Automation
playwright==1.42.0

# Rust Integration & High-Performance Extensions
maturin==1.5.0

# Encryption
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email
aiosmtplib==3.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
